Add possibility to explicitly signal root directory
When a root directory is not directly specified in the config, TexLab
traverses upwards in the file system hierarchy upwards until a .git
directory or a latexmk/chktex/Tectonic configuration file is found. The
first directory where one of the search items is found is then
considered the root directory. This is a reasonable default but may not
work for all project structures.

To provide more flexibility, this commit adds the possibility to
explicitly signal when TexLab should stop searching for the root
directory via empty ".texlabroot" or "texlabroot" files. These serve no
purpose other than signaling to TexLab that it should stop traversing
upwards at this point.