Initial work towards prefixes for custom label commands.
This commit mainly extends the configs with a map from label command
name to the respective prefix.
It does not yet implement any functionality, since it appears that
`base-db/semantics/label` needs an extension carrying a span that
contains the label command.
The plan is to extend the logic in `crates/diagnostics/src/labels.rs`.