Consider prefixes when renaming.
This commit implements key functionality for renaming labels with prefixes.
A sane assumption it does is that all labels found as candidate for
renaming share a common prefix, up to looking it up first and prepending it, e.g., for `\goal{foo}`.
Instead of storing a renaming candidate for each entry, it only keeps
track of the prefixes and prepends them accordingly, depending on the
renaming candidate, by swapping the `TextRange` with `RenameInformation` inside the `RenameResult`.
Unfortunately, this pollutes a bit the other renaming ops, such as
commands or citations, which don't have prefixes.
Nevertheless, changes there have been incremental and `RenameInformation` implements a `From<TextRange>` to easily map an existing `TextRange` into it, simply assuming an empty prefix.
In terms of tests, the `prefix` should be ignored.