Add configurable prefixes for custom label commands. (#1139)
This commit extends the configs with a map from label command name to the respective prefix:
*  Prepends user-defined prefixes to labels in the semantics
pass. Note that this yields to a mismatch between the actual length of a
span and its range, but there are no (anticipated) bad side-effects due
to this.
* Adds two tests that ensure completion for custom labels with custom
prefix works as expected. The `\ref` command should list the prefix,
while a custom reference command with a configured prefix should not.
* Implements key functionality for renaming labels with prefixes.
A sane assumption it does is that all labels found as candidate for
renaming share a common prefix, up to looking it up first and prepending it, e.g., for `\goal{foo}`.
Instead of storing a renaming candidate for each entry, it only keeps
track of the prefixes and prepends them accordingly, depending on the
renaming candidate, by swapping the `TextRange` with `RenameInformation` inside the `RenameResult`.
Unfortunately, this pollutes a bit the other renaming ops, such as
commands or citations, which don't have prefixes.
Nevertheless, changes there have been incremental and `RenameInformation` implements a `From<TextRange>` to easily map an existing `TextRange` into it, simply assuming an empty prefix.
In terms of tests, the `prefix` should be ignored.