Add latexindent replacement options (#1241)
Add an option field "replacement" to the LatexIndentConfig structs, and modify the build_arguments function to check if it config.replacement contains one of "-r", "-rr", or "-rv". If so, it appends that to the list of arguments to LatexIndent. If it does not exist, or it contains any other string, the option is ignored.