Improve error message when build.executable doesn't exist
texlab outputs on standard error:

	ERROR - Failed to compile document "file:///path/to/test.tex":
	Unable to run compiler: No such file or directory (os error 2)

which is confusing because the actual issue is that "latexmk" was
not found.  Fix that by adding the program name.  This might not be
as appropriate for all possible io::Error kinds but it's probably
useful for a few of them.

Also we should remove a file:// prefix when printing URLs,
but that's a separate change.