Capabilities for LSP extensions
Clients sending textDocument/build to a server that doesn't support
it may receive an error.  In a multi-server setup, users expect to
send the message only to servers that support it.  Add a capability
to enable this use case.

To-do: add this to https://github.com/latex-lsp/texlab/wiki/LSP-Internals

Closes #1328